Lmoares b11

+++ ContractDevApi/ContractDevApi/Controllers/UserAccountsController.cs
@@ -135,14 +135,14 @@ namespace ContractDevApi.Controllers
 
             _context.SocialConnections.Add(socials);
 
-            var reviews = new UserReview
-            {
-                UserAccountId = user.UserAccountId,
-                NumberOfReviews = 0,
-                TotalReviewPoints = 0.0f,
-                AverageReviewScore = 0.0f,
-                UserAccount = user
-            };
+            //var reviews = new UserReview
+            //{
+            //    UserAccountId = user.UserAccountId,
+            //    NumberOfReviews = 0,
+            //    TotalReviewPoints = 0.0f,
+            //    AverageReviewScore = 0.0f,
+            //    UserAccount = user
+            //};
 
             //Try to update database -- if unsuccessful return error
             try {
@@ -319,6 +319,10 @@ namespace ContractDevApi.Controllers
 
             //If password does not pass verification, return error Status 401
             if (!validatePassword) return Unauthorized(new { Message = "Invalid Password" });
+
+            bool validateSecurityAnswer = BCrypt.Net.BCrypt.Verify(dto.SecurityAnswer.ToLower(), user.SecurityAnswer);
+
+            if (!validateSecurityAnswer) return Unauthorized(new { Message = "Invalid Security Answer" });
 
 
             _context.UserAccounts.Remove(userAccount);
